misc (#219)
* update xattree and modify the array field decorator wrappers to accommodate a new dtype arg, this gives us more control over dtype than the type hint. the actual arg overrides the type hint, but the type hint is used if no arg is passed
* make `boundname` arrays on various components a fixed length string of length 40 (`LENBOUNDNAME`)
* devtools sorts blocks so no need to move options to top in converter
* switch oc arrays from object -> str dtype
* turn type checking off for now in vscode settings
* add pytest-cov as a test dependency
* leave some todo comments

+++ flopy4/mf6/gwf/dis.py
@@ -64,7 +64,7 @@ class Dis(Package):
         dims=("nlay", "nrow", "ncol"),
         converter=Converter(dict_to_array, takes_self=True, takes_field=True),
     )
-    idomain: Optional[NDArray[np.int32]] = array(
+    idomain: Optional[NDArray[np.int64]] = array(
         block="griddata",
         default=1,
         dims=("nlay", "nrow", "ncol"),
